Description
Change the value of x to a stock ticker symbol to analyze. 

The Aroon indicator is a technical indicator that is used to identify trend changes in the price of an asset, as well as the strength of that trend. In essence, the indicator measures the time between highs and the time between lows over a time period. The idea is that strong uptrends will regularly see new highs, and strong downtrends will regularly see new lows. The indicator signals when this is happening, and when it isn't. https://www.investopedia.com/terms/a/aroon.asp#:~:text=advisor%20(CTA).-,What%20Is%20the%20Aroon%20Indicator%3F,lows%20over%20a%20time%20period.

The accumulation/distribution indicator (A/D) is a cumulative indicator that uses volume and price to assess whether a stock is being accumulated or distributed. The A/D measure seeks to identify divergences between the stock price and the volume flow. This provides insight into how strong a trend is. https://www.investopedia.com/terms/a/accumulationdistribution.asp#:~:text=The%20accumulation%2Fdistribution%20indicator%20(A%2FD)%20is%20a,how%20strong%20a%20trend%20is.